Structural improvements and code changes to login screen components
